+++ studio-project/src/App.jsx
@@ -3,13 +3,13 @@ import Graph from "./graph.jsx";
 import "./index.css";
 import Chat from "./chat.jsx";
 import './App.css'
-import { ReactFlow, ReactFlowProvider } from "@xyflow/react";
-import { useDataToGraphTransformer } from "./hooks/useGraphTransformer.js";
+import { ReactFlowProvider } from "@xyflow/react";
+import { transformKnowledgeGraph } from "./data/knowledgeGraphData.js";
 
 
-function App() {
+export default function App() {
   const [leftWidth, setLeftWidth] = useState(66.6); // start ~2/3
-  const data = useDataToGraphTransformer();
+  const data = transformKnowledgeGraph();
   const containerRef = useRef(null);
 
   const handleMouseDown = (e) => {
@@ -37,6 +37,7 @@ function App() {
 
   return (
     <div ref={containerRef} className="flex h-screen w-screen">
+      
       <div
         className="h-full bg-gray-100 overflow-hidden "
         style={{ width: `${leftWidth}%` }}
@@ -45,7 +46,7 @@ function App() {
         <Graph data= {data} width={100-leftWidth} />
         </ReactFlowProvider>
       </div>
-
+      
       <div
         className="w-1 bg-gray-400 cursor-col-resize hover:bg-gray-600"
         onMouseDown={handleMouseDown}
@@ -58,4 +59,3 @@ function App() {
   );
 }
 
-export default App;
