+++ studio-project/src/data/atoms.js
@@ -1,4 +1,4 @@
-import { atom, selectAtom } from "jotai";
+import { atom } from "jotai";
 
 //Atom for the chat Component
 export const messagesAtom = atom([
@@ -15,19 +15,8 @@ export const textStatusAtom = atom('ready');
 export const textAtom = atom('');
 export const timerAtom = atom(null);
 
-//Atom for the knowledge graph component
-export const graphDataAtom = atom({
-    nodes: [],
-    edges: [],
-    draggingNodeId:null,
-    selectedNodeId:null,
-
-});
-
-export const nodesAtom = selectAtom(graphDataAtom, (u) => u?.nodes ?? []);
-
-export const edgesAtom = selectAtom(graphDataAtom, (u) => u?.edges ?? []);
-
-export const selectedNodeAtom = selectAtom(graphDataAtom, (u) => u?.selectedNodeAtom ?? null);
-
-export const draggingNodeIdAtom = selectAtom(graphDataAtom, (u) => u?.draggingNodeId ?? null);
+//Atoms for the graph component
+export const nodesAtom = atom([]);
+export const edgesAtom = atom([]);
+export const selectedNodeAtom = atom(null);
+export const draggingNodeIdAtom = atom(null);
